Add files via upload
The 8-puzzle problem is a puzzle invented and popularized by Noyes Palmer Chapman in the 1870s. It is
played on a 3-by-3 grid with 8 square blocks labeled 1 through 8 and a blank square. Your goal is to
rearrange the blocks so that they are in order. You are permitted to slide blocks horizontally or vertically
into the blank square.
In this repo, we solve it using two algorithms 
1. Hill climbing.
2. A* search